+++ demisto_sdk/commands/content_graph/tests/create_content_graph_test.py
@@ -410,7 +410,10 @@ class TestCreateContentGraph:
             # make sure that the extracted files are all .csv
             extracted_files = list(tmp_path.glob("extracted/*"))
             assert extracted_files
-            assert all(file.suffix == ".csv" for file in extracted_files)
+            assert all(
+                file.suffix == ".csv" or file.name == "metadata.json"
+                for file in extracted_files
+            )
 
     def test_create_content_graph_relationships(
         self,