+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 * Changed **validate** to allow hiding parameters of type 0, 4, 12 and 14 when replacing with type 9 (credentials) with the same name.
 * Fixed an issue where **update-release-notes** fails to update *MicrosoftApiModule* dependent integrations.
 * Fixed an issue where the **upload** command failed because `docker_native_image_config.json` file could not be found.
-* Updated the **validate** and **update-release-notes** commands to unskip the *Triggers Recommendations* content type.
+* Added a metadata file to the content graph zip, to be used in the **update-content-graph** command.
+
+
 
 ## 1.8.2
 * Fixed an issue where demisto-py failed to upload content to XSIAM when `DEMISTO_USERNAME` environment variable is set.